drf_project/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==36.0.2
Django==4.0.4
django-filter==21.1
django-rest-knox==4.2.0
django-rest-passwordreset==1.2.1
django-shortcuts==1.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docopt==0.6.2
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
